requirements.txt
streamlit==1.22.0
PyPDF2==3.0.1
python-docx==0.8.11
spacy==3.5.0
scikit-learn==1.2.0
pandas==1.5.3
plotly==5.11.0